Organize classes into different packages
