+++ src/main/java/TrackBot/commands/ListCommand.java
@@ -2,8 +2,8 @@ package TrackBot.commands;
 
 import TrackBot.TrackBotException;
 import TrackBot.TrackBotStorage;
-import TrackBot.TrackList;
-import TrackBot.Ui;
+import TrackBot.task.TrackList;
+import TrackBot.ui.Ui;
 
 
 public class ListCommand extends Command {
