+++ src/main/java/TrackBot/commands/MarkCommand.java
@@ -2,8 +2,8 @@ package TrackBot.commands;
 
 import TrackBot.TrackBotException;
 import TrackBot.TrackBotStorage;
-import TrackBot.TrackList;
-import TrackBot.Ui;
+import TrackBot.task.TrackList;
+import TrackBot.ui.Ui;
 
 public class MarkCommand extends Command {
     private final int taskIndex;
