+++ src/main/java/TrackBot/ui/Ui.java
@@ -1,4 +1,4 @@
-package TrackBot;
+package TrackBot.ui;
 
 public class Ui {
     public void showWelcome() {
